pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "galactic-wrestling"
version = "0.1.0"
description = "Galactic Wrestling — terminal deckbuilder (ASCII)"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "textual>=0.86.0",
    "argon2-cffi>=23.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
galactic-wrestling = "galactic_wrestling.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/galactic_wrestling"]

[tool.hatch.build.targets.sdist]
include = ["src/galactic_wrestling"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-ra",
    "--strict-markers",
    "--import-mode=importlib",
    "--cov=galactic_wrestling",
    "--cov-branch",
    "--cov-report=term-missing",
    "--cov-fail-under=100",
]

[tool.coverage.run]
source_pkgs = ["galactic_wrestling"]
# Textual screens are thin glue; behaviour is exercised via smoke tests and manual QA.
omit = [
    "*/galactic_wrestling/__main__.py",
    "*/galactic_wrestling/match/__main__.py",
    "*/galactic_wrestling/ui/screens.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]
